fix(errors): rename errorsx
