+++ internal/errorsx/errorsx.go
@@ -1,4 +1,4 @@
-package errors
+package errorsx
 
 import "errors"
