+++ internal/errorsx/missing.go
@@ -1,4 +1,4 @@
-package errors
+package errorsx
 
 import "fmt"
