+++ internal/helpers/helper.go
@@ -1,21 +1,21 @@
 package helpers
 
-import "github.com/shipengqi/commitizen/internal/errors"
+import "github.com/shipengqi/commitizen/internal/errorsx"
 
-func GetValueFromYAML[T any](data map[string]interface{}, key string) (T, error) {
+func GetValueFromYAML[T any](data map[string]any, key string) (T, error) {
 	var (
 		res T
 		ok  bool
-		v   interface{}
+		v   any
 	)
 
 	v, ok = data[key]
 	if !ok {
-		return res, errors.NewMissingErr(key)
+		return res, errorsx.NewMissingErr(key)
 	}
 	res, ok = v.(T)
 	if !ok {
-		return res, errors.ErrType
+		return res, errorsx.ErrType
 	}
 	return res, nil
 }
