+++ internal/parameter/validators/slice.go
@@ -1,11 +1,11 @@
 package validators
 
-import "github.com/shipengqi/commitizen/internal/errors"
+import "github.com/shipengqi/commitizen/internal/errorsx"
 
 func MultiRequired(name string) func([]string) error {
 	return func(vals []string) error {
 		if len(vals) == 0 {
-			return errors.NewRequiredErr(name)
+			return errorsx.NewRequiredErr(name)
 		}
 		return nil
 	}
